Rewrite entryOrGUID, int64 again but changed variable for loading checks uint32 *clarify variable namings

+++ src/game/CreatureEventAI.h
@@ -434,7 +434,7 @@ struct CreatureEventAI_Action
 struct CreatureEventAI_Event
 {
     uint32 event_id;
-    uint32 entryOrGUID; // needs to be uint32 for correct work
+    int64 entryOrGUID; // needs to be int64 for correct work
     uint32 event_inverse_phase_mask;
 
     EventAI_Type event_type : 16;
@@ -575,7 +575,7 @@ struct CreatureEventAI_Event
 
 //Event_Map
 typedef std::vector<CreatureEventAI_Event> CreatureEventAI_Event_Vec;
-typedef UNORDERED_MAP<int64, CreatureEventAI_Event_Vec > CreatureEventAI_Event_Map;
+typedef UNORDERED_MAP<int64, CreatureEventAI_Event_Vec> CreatureEventAI_Event_Map;
 
 struct CreatureEventAI_Summon
 {
